requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.5
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-filter==24.3
djangorestframework==3.15.2
PyJWT==2.10.1
pytz==2025.1
six==1.17.0
sqlparse==0.5.1
tzdata==2024.2
